+++ packages/dds/tree/src/test/shared-tree/fuzz/fuzzUtils.ts
@@ -13,6 +13,7 @@ import type { IIdCompressor, SessionId } from "@fluidframework/id-compressor";
 import {
 	createIdCompressor,
 	deserializeIdCompressor,
+	toIdCompressorWithCore,
 	type IIdCompressorCore,
 } from "@fluidframework/id-compressor/internal";
 
@@ -256,11 +257,13 @@ export const createOrDeserializeCompressor = (
 	sessionId: SessionId,
 	summary?: FuzzSerializedIdCompressor,
 ): IIdCompressor & IIdCompressorCore => {
-	return summary === undefined
-		? createIdCompressor(sessionId)
-		: summary.withSession
-			? deserializeIdCompressor(summary.serializedCompressor)
-			: deserializeIdCompressor(summary.serializedCompressor, sessionId);
+	return toIdCompressorWithCore(
+		summary === undefined
+			? createIdCompressor(sessionId)
+			: summary.withSession
+				? deserializeIdCompressor(summary.serializedCompressor)
+				: deserializeIdCompressor(summary.serializedCompressor, sessionId),
+	);
 };
 
 export const deterministicIdCompressorFactory: (